refactor: rename component and style names
